+++ src/libs/extensionsystem/invoker.cpp
@@ -69,7 +69,7 @@ void InvokerBase::invoke(QObject *t, const char *slot)
         if (paramCount)
             sig.append(',');
         const char *type = arg[paramCount].name();
-        sig.append(type, strlen(type));
+        sig.append(type, static_cast<int>(strlen(type)));
     }
     sig.append(')');
     sig.append('\0');
